人工智能基础/yolov8_helmet_detection/ui/requirements.txt
numpy==1.24.3
opencv_python==4.8.1.78
opencv_python_headless==4.7.0.72
Pillow==10.1.0
PySide6==6.6.1
PySide6_Addons==6.6.1
PySide6_Essentials==6.6.1
shiboken6==6.6.1
torch==2.0.1
ultralytics==8.0.227
